Add files via upload
This is my most complex project to date with one major issue being how I handle the recover times. I will comment everything in the coming days. But for now just know I had to use a periodic timer for each person to enable pausing where people ceased recovering. It is my understanding that once starting a one-time java.util.Timer with no periodic repetition it can not be stopped the way I wished to pause my animation. This is also my first exposure to volatile types and the keyword synchronized. All of the source code is here as well as a runnable .jar file if you do not wish to download every file and you just wish to see the result. I am also kind of new to saving and loading files, so if a folder is not created on your desktop please try to educate me as to going about fixing that issue. Hope you enjoy. Please give constructive criticism.